Added Databse Files And A New Analyze
added database related files to store the analysis that we make in the spark_analysis.py file. databse has one table for the new analyze that we added to the analysis file. we also added a configuration file to not to hard code the password and other important thing to connect to the database