Added server.py To Handle API Calls
added server.py to the project for the backend part of the project to response to the API calls that we are going to make from the frontend part in the future. added a simple login and a login route. also added a generate secret key script to make a seceret key for the authentication. we used flask for the backend