Added All The Other Components For Other Analysis
added 4 more components for the four other analysis that we have. in the components we handle the pagination and also search by IP for the user to have a better experience. in the future we will add two more components to add charts to two of the more important analysis means IPCluster and Anomaly detection